F Keyboarding
given a virtual keyboard layout and your task is to determine the
minimal number of strokes needed to type a given text, where pressing
any of the ve hardware
buttons constitutes a stroke. The keys are arranged in a rectangular
grid, such that each virtual
key occupies one or more connected unit squares of the grid. The cursor
starts in the upper left
corner of the keyboard and moves in the four cardinal directions, in
such a way that it always skips
to the next unit square in that direction that belongs to a dierent key.
If there is no such unit
square, the cursor does not move.

The rst line of the input contains two integers r and c (1  r; c 50),
giving the number of rows
and columns of the virtual keyboard grid. The virtual keyboard is
specied in the next r lines, each
of which contains c characters. The possible values of these characters
are uppercase letters, digits,a dash, and an asterisk (representing
Enter). There is only one key corresponding to any given
character. Each key is made up of one or more grid squares, which will
always form a connected
region. The last line of the input contains the text to be typed. This
text is a non-empty string of
at most 10 000 of the available characters other than the asterisk.